requirements.txt
gym
dotmap
carbontracker
neptune_client
matplotlib
scipy
torch
mlagents_envs
gym_unity
seaborn
unityagents